core/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
beyonic==0.1.16
billiard==4.1.0
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
deprecation==2.1.0
dj-database-url==2.1.0
django-ace==1.26.0
django-autocomplete-light==3.9.7
django-codemirror2==0.2
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-cte==1.1.4
django-currentuser==0.6.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-js-asset==2.1.0
django-json-widget==1.1.1
django-mathfilters==1.0.0
django-model-utils==4.3.1
django-mptt==0.9.1
django-notifications-hq==1.8.3
django-phonenumber-field==7.2.0
django-qr-code==3.1.1
django-tinymce==3.6.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dry-rest-permissions==0.1.10
ecdsa==0.18.0
et-xmlfile==1.1.0
future==0.18.3
gunicorn==21.2.0
idna==3.4
jsonfield==3.1.0
kombu==5.3.2
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
phonenumbers==8.13.24
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3.post1
pyxlsb==1.0.10
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
segno==1.5.3
sentry-sdk==1.34.0
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
